fix: rename folder with camelCase
